web/requirements.txt
Django==1.11.3
django-leaflet==0.22.0
django-cors-headers==2.1.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
psycopg2==2.7.2
pytz==2017.2
